LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-link-element/document-without-browsing-context-expected.txt


FAIL Create a document, adopt the node assert_equals: expected "1" but got "0"
FAIL Create a stylesheet in innerHTML document assert_equals: expected "1" but got "0"
